requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
exceptiongroup==1.2.1
fastapi==0.110.2
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python-headless==4.9.0.80
orjson==3.10.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
